refactor: rename component, api and add changelogs
