+++ changelogs/unreleased/76021.json
@@ -1,0 +1,5 @@
+{
+  "title": "AopChart: component that will display Aop graphics",
+  "type": "feat",
+  "packages": "ui"
+}
